+++ src/workflow/__init__.py
@@ -2,7 +2,7 @@
 prompt-chaining workflow
 """
 
-__version__ = "0.4.5"
+__version__ = "0.4.6"
 __author__ = "Christopher Scragg"
 __email__ = "clscragg@protonmail.com"